backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pypdf>=3.17.1
python-multipart>=0.0.6
requests>=2.31.0
google-generativeai>=0.3.1
